chinh sua db

+++ src/migrations/create-post.js
@@ -34,10 +34,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       priceNumber: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       areaNumber: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       provinceCode: {
         type: Sequelize.STRING
